Create README.md
In this project, we are going to build a Fingerprint Door Lock using Arduino Nano with Bluetooth Module, Relay Module, Solenoid Lock, and Android SmartPhone. An Android application is used, which scans and verifies the fingerprint and sends confirmation data to Arduino through Bluetooth serial communication. This system uses a built-in smartphone Fingerprint sensor to lock and unlock the door.